requirements.txt
python-dotenv
langchain
langchain-chroma
langchain_community
langchain-huggingface